refactor: create features/layout-library directory (Phase 5e)
Move layout manager components and hooks to new feature directory:
- src/components/Modals/LayoutManagerModal/ → src/features/layout-library/components/LayoutManagerModal/
- src/hooks/useLayoutSwitcher.ts → src/features/layout-library/hooks/
- src/hooks/useLayoutRouting.ts → src/features/layout-library/hooks/

Create barrel export at src/features/layout-library/index.ts

Add re-export stubs at old locations for backward compatibility:
- src/components/Modals/LayoutManagerModal/index.ts
- src/hooks/useLayoutSwitcher.ts
- src/hooks/useLayoutRouting.ts

Update test imports to use new paths.
