+++ src/components/Modals/LayoutManagerModal/index.ts
@@ -1,0 +1,2 @@
+// Re-export from new location for backward compatibility
+export { LayoutManagerModal } from '../../../features/layout-library/components/LayoutManagerModal';
