+++ src/features/layout-library/components/LayoutManagerModal/ImportView.tsx
@@ -1,8 +1,8 @@
 import { useState, useRef, useCallback } from 'react';
 import type { ChangeEvent, DragEvent } from 'react';
-import { validateImport } from '../../../shared/utils/validation';
-import { decodeLayoutFromURL } from '../../../core/storage';
-import type { Layout } from '../../../core/types';
+import { validateImport } from '../../../../shared/utils/validation';
+import { decodeLayoutFromURL } from '../../../../core/storage';
+import type { Layout } from '../../../../core/types';
 
 interface ImportViewProps {
   onImport: (layout: Layout) => void;
