+++ src/features/layout-library/components/LayoutManagerModal/LayoutList.tsx
@@ -1,9 +1,9 @@
 import { useState, useRef, useCallback } from 'react';
-import type { LayoutEntry, Layout } from '../../../core/types';
+import type { LayoutEntry, Layout } from '../../../../core/types';
 import { LayoutListItem } from './LayoutListItem';
-import { useLayoutStore } from '../../../core/store/layout';
-import { loadLayoutByIdAsync, downloadLayoutAsFile } from '../../../core/storage';
-import { useUIStore } from '../../../core/store/ui';
+import { useLayoutStore } from '../../../../core/store/layout';
+import { loadLayoutByIdAsync, downloadLayoutAsFile } from '../../../../core/storage';
+import { useUIStore } from '../../../../core/store/ui';
 
 /** Threshold for showing search bar */
 const SEARCH_THRESHOLD = 6;
