+++ src/features/layout-library/components/LayoutManagerModal/SharedWithMeItem.tsx
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
-import type { SharedWithMeEntry } from '../../../core/types';
-import { LayoutThumbnail } from '../../LayoutThumbnail';
+import type { SharedWithMeEntry } from '../../../../core/types';
+import { LayoutThumbnail } from '../../../../components/LayoutThumbnail';
 
 interface SharedWithMeItemProps {
   entry: SharedWithMeEntry;
