+++ src/features/layout-library/index.ts
@@ -1,0 +1,6 @@
+// Components
+export { LayoutManagerModal } from './components/LayoutManagerModal';
+
+// Hooks
+export { useLayoutSwitcher } from './hooks/useLayoutSwitcher';
+export { useLayoutRouting } from './hooks/useLayoutRouting';
