+++ src/test/components/ImportView.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { render, screen, fireEvent, cleanup, waitFor } from '@testing-library/react';
-import { ImportView } from '../../components/Modals/LayoutManagerModal/ImportView';
+import { ImportView } from '../../features/layout-library/components/LayoutManagerModal/ImportView';
 import * as validation from '../../shared/utils/validation';
 import * as storage from '../../core/storage';
