+++ src/test/components/LayoutListItem.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen, fireEvent } from '@testing-library/react';
-import { LayoutListItem } from '../../components/Modals/LayoutManagerModal/LayoutListItem';
+import { LayoutListItem } from '../../features/layout-library/components/LayoutManagerModal/LayoutListItem';
 import type { LayoutEntry } from '../../core/types';
 
 // Mock LayoutThumbnail
